Separated Database Classes + Ryan Lake
Added Ryan Lake to players.
Removed Database Classes from main activity.
